Add files via upload
Online food ordering system in PHP / MySQL
This Online Food Ordering System using PHP and MySQL has both the admin side and visitor/client side. The admin will manage all the data needed for the website and the list of available menus. The client will browse the website, explore and choose his/her desired menu’s and save it from a cart for later checkout. After that, the client will go to the cart, review his/ her order, confirm the delivery address and other information, and place his/ her order. This is an easy-to-use project. This project will be a good start if you’re planning to build or develop an ordering system for a restaurant or cafe. 
Administrator logs ;
Username: admin
Password: admin123